Extracted utility function to create menu entries.
